moved label fetcher to its own package
