+++ apps/dataset-browser/package.json
@@ -21,19 +21,17 @@
   },
   "dependencies": {
     "@colonial-collections/iris": "*",
+    "@colonial-collections/label-fetcher": "*",
     "@floating-ui/react": "0.23.0",
     "@hapi/hoek": "11.0.2",
     "@headlessui/react": "1.7.14",
     "@heroicons/react": "2.0.17",
     "@next/mdx": "13.3.0",
     "classnames": "2.3.2",
-    "fetch-sparql-endpoint": "3.2.1",
-    "lru-cache": "9.0.1",
     "next": "13.2.4    ",
     "next-intl": "2.13.0-beta.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "validate-iri": "1.0.1",
     "zod": "3.21.4"
   },
   "devDependencies": {
@@ -47,6 +45,7 @@
     "cypress": "12.9.0",
     "encoding": "0.1.13",
     "eslint-config-custom": "*",
+    "fetch-sparql-endpoint": "3.2.1",
     "gts": "3.1.1",
     "jest": "29.5.0",
     "postcss": "8.4.22",
