+++ packages/label-fetcher/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "@colonial-collections/label-fetcher",
+  "version": "1.0.0",
+  "private": true,
+  "engines": {
+    "node": "18.x"
+  },
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "MIT",
+  "scripts": {
+    "clean": "gts clean",
+    "compile": "tsc",
+    "fix": "gts fix",
+    "lint": "gts check",
+    "test": "jest --passWithNoTests --testPathIgnorePatterns '<rootDir>/src/.*/*.integration.test.ts'",
+    "test:ci": "npm test -- --ci",
+    "test:watch": "npm test -- --watchAll",
+    "test:integration": "jest --testMatch '**/src/**/*.integration.test.ts'",
+    "test:integration:watch": "npm run test:integration -- --watchAll"
+  },
+  "dependencies": {
+    "@colonial-collections/iris": "*",
+    "fetch-sparql-endpoint": "3.2.1",
+    "lru-cache": "9.0.1",
+    "zod": "3.21.4"
+  },
+  "devDependencies": {
+    "@jest/globals": "29.5.0",
+    "@types/jest": "29.5.0",
+    "@types/node": "18.15.11",
+    "eslint-config-custom": "*",
+    "gts": "3.1.1",
+    "jest": "29.5.0",
+    "ts-jest": "29.1.0",
+    "tsconfig": "*",
+    "typescript": "5.0.4"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testTimeout": 60000,
+    "testMatch": ["<rootDir>/src/**/*.test.ts"],
+    "collectCoverage": true,
+    "collectCoverageFrom": ["<rootDir>/src/**/*.ts"]
+  }
+}
