+++ packages/label-fetcher/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "tsconfig/base.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./build",
+    "noEmitOnError": false
+  },
+  "include": ["src/**/*.ts"]
+}
